requirements.txt
Django==3.0.2
django-nose==1.4.6
djangorestframework==3.11.0
nose==1.3.7
Pillow==7.0.0
sqlparse==0.3.0
